[refactor] Extract assertion type predicates to ``_guards.py``
